server/requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.1
greenlet==0.4.15
gunicorn==19.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
monotonic==1.5
pycparser==2.19
PyJWT==1.7.1
pymongo==3.10.0
PyMySQL==0.9.3
python-engineio==3.11.1
python-socketio==4.4.0
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.7
Werkzeug==0.16.0
